+++ cpp/prac3.cpp
@@ -1,10 +1,10 @@
-#include <iostrem>
+#include <iostream>
 #include <cmath>
 
 int main(void) {
     int R, H;
     std::cin >> R >> H;
-    int V{ M_PI * R * R * H };
+    double V{ M_PI * (R * R * H) };
     std::cout << V << std::endl;
     return 0;
 }
